requirements.txt
pdfplumber==0.11.5
pytesseract==0.3.13
Pillow==11.2.1
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
python-docx==1.1.2
sentence-transformers==3.4.1
psycopg2-binary==2.9.10
requests==2.32.3
redis==5.2.1
python-dotenv==1.0.1
